need t introduce type casting

+++ area-circle.c
@@ -4,17 +4,17 @@ float pi = 3.14;
 
 
 int area(int diameter) {
-         int radius = diameter /2 ;
-         int areaOfCircle = 0;
+         float radius = diameter /2.0 ;
+         float areaOfCircle = 0;
         areaOfCircle =  pi * (radius * radius);
 
         return areaOfCircle;
 }
 
 int main () {
-    int diameter = 0;
+    float diameter = 0;
     printf("Please enter the diameter: ");
-    scanf("%d", &diameter);
+    scanf("%f", &diameter);
 
     float final =  area(diameter);
     printf("Area: %.2f", final );
